Phase 0 WP-G1b: de-Tauri whisper models, speaker diarization services, and LLM provider clients (ollama/anthropic/groq/openai/openrouter)
Splits each mixed file in this work package's area into a pure-core file
(state, HTTP clients, DB-backed business logic) and a pure-shell
commands.rs that only wraps #[tauri::command] fns around the core calls:

- whisper_engine/commands.rs -> new whisper_engine/models.rs (WHISPER_ENGINE
  static, model dir, download/validate logic) + thin command wrappers,
  re-exported from commands.rs so existing
  `whisper_engine::commands::WHISPER_ENGINE`-style paths keep working.
- speaker_diarization/commands.rs -> new speaker_diarization/service.rs
  (model download/build/refine/promote/merge core logic); commands.rs
  re-exports build_diarizer/try_init_for_recording/refine_and_persist/
  shutdown_for_recording. audio::import's ensure_pyannote_segmentation_model
  call now points at the core service fn instead of the command.
- ollama: ollama.rs -> client.rs (pure core), new commands.rs with thin
  #[tauri::command] wrappers.
- anthropic/groq/openai/openrouter: each provider file had exactly one
  #[tauri::command]-decorated model-listing fn; moved it into a new
  commands.rs per provider, leaving the client/parsing code Tauri-free.
- summary/summary_engine/commands.rs -> new service.rs (ModelManagerState,
  init_model_manager, download/recommend/available-model logic, an
  ensure_manager helper shared by the command wrappers).

lib.rs generate_handler! paths updated for every moved command
(ollama::commands::*, anthropic::commands::*, groq::commands::*,
openai::commands::*, openrouter::commands::*).

calendar/commands.rs, summary/commands.rs, summary/template_commands.rs,
and src/api/**, src/notifications/** were already pure-shell (or already
correctly split) and needed no changes; summary/templates/mod.rs's one
"tauri" grep hit is a doc-comment path reference, not a real dependency.

Zero new cargo check warnings; 300/300 lib tests pass.
