+++ frontend/src-tauri/src/anthropic/mod.rs
@@ -1,1 +1,2 @@
 pub mod anthropic;
+pub mod commands;
