+++ frontend/src-tauri/src/groq/mod.rs
@@ -1,1 +1,2 @@
+pub mod commands;
 pub mod groq;
