+++ frontend/src-tauri/src/lib.rs
@@ -825,13 +825,13 @@ pub fn run() {
             // replaces the old IndexedDB scan)
             audio::recovery_commands::list_interrupted_meetings,
             audio::recovery_commands::recover_meeting,
-            ollama::get_ollama_models,
-            ollama::pull_ollama_model,
-            ollama::delete_ollama_model,
-            ollama::get_ollama_model_context,
-            openai::openai::get_openai_models,
-            anthropic::anthropic::get_anthropic_models,
-            groq::groq::get_groq_models,
+            ollama::commands::get_ollama_models,
+            ollama::commands::pull_ollama_model,
+            ollama::commands::delete_ollama_model,
+            ollama::commands::get_ollama_model_context,
+            openai::commands::get_openai_models,
+            anthropic::commands::get_anthropic_models,
+            groq::commands::get_groq_models,
             api::api_get_meetings,
             api::api_search_transcripts,
             api::api_get_model_config,
@@ -891,7 +891,7 @@ pub fn run() {
             summary::summary_engine::commands::builtin_ai_is_model_ready,
             summary::summary_engine::commands::builtin_ai_get_available_summary_model,
             summary::summary_engine::commands::builtin_ai_get_recommended_model,
-            openrouter::get_openrouter_models,
+            openrouter::commands::get_openrouter_models,
             audio::recording_preferences::get_recording_preferences,
             audio::recording_preferences::set_recording_preferences,
             audio::recording_preferences::get_default_recordings_folder_path,
