+++ frontend/src-tauri/src/ollama/mod.rs
@@ -1,6 +1,6 @@
+pub mod client;
 pub mod commands;
 pub mod metadata;
-pub mod ollama;
 
-pub use ollama::*;
+pub use client::*;
 // Don't re-export commands to avoid conflicts - lib.rs will import directly
