+++ frontend/src-tauri/src/openai/mod.rs
@@ -1,1 +1,2 @@
+pub mod commands;
 pub mod openai;
