+++ frontend/src-tauri/src/whisper_engine/mod.rs
@@ -1,5 +1,6 @@
 pub mod commands;
 pub mod lease;
+pub mod models;
 pub mod whisper_engine;
 // pub mod stderr_suppressor;
